Add support for reading input maps
Okay, this is a massive commit (and I feel bad about it).

This commit splits PrescribedDataStatic and PrescibedDataTemporal in
multiple independent modules.

Before, PrescribedData was:
- reading files
- regridding
- discovering dates available
- interpolating
- taking care of edge cases (e.g., what to do when reading outside of
where the date is outside of the definition range)

This commit splits all the different functions in different independent
modules so that they can be more easily maintained and upgraded.

This commit also introduces InterpolationsRemapper, which uses
Interpolations.jl to do online/ and distributed remapping from files.

This will soon be needed to efficiently support reading and buffering
reads for GPU runs.