Add a tutorial of a single site perfect model calibration
This commit a calibration tutorial, with detailed
explanations. We calibrate two parameters, g1 and g0, on
latent heat flux data target, using EnsembleKalmanProcesses.jl.
We use a single site model, at the ozark site, and calibrate first on
simulation output, with the goal of demonstrating the ability to
recover the original parameters and demonstrate a simple calibration
framework.