Add ClimaLand.Artifacts
The Julia Artifacts system has some deficiencies. For example, it does
not stack `Artifacts.toml` files. This is a problem, especially with
`AritfactsWrapper` around, which adds new `Artifacts.toml`. This PR
establishes ownership of artifacts in ClimaLand itself. This ensures
that the `Artifacts.toml` being used is the main one.