Add @with_error, remove @generate_error_functions
`@generate_error_functions` was not a great macro: it required adding
new names to a line every time a diagnostic is added, it was dealing
with several variables at the same time, and could not be used outside
of its scope of definition and/or called twice.

`@with_error` accomplishes a the same goals as
`@generate_error_functions` without the problems. The only difference is
that now `compute` functions have to be decorated.