Dict for single site runs
Add diagnostic_as_vectors

`diagnostic_as_vectors` is a utility function to use ClimaDiagnostics with
DictWriter for column simulations. Assuming `writer` is a writer, this is
called as

`diagnostic_as_vectors(writer, "gpp_1d_average")`

and returns two vectors, the times, and the values.

This works only on CPU and for PointSpaces