Change PlantHydraulicsParameters to hold rooting_depth
PlantHydraulicsParameters previously held a root distribution function
that had a default value for rooting depth. This function is
always the same outside testing, so it was removed as a parameter
and placed into src. rooting_depth, an argument to the distribution function
is now a parameters that is either a Float or a field. The tests are changed to
check passing rooting_depth as a float or a field of floats that are the same everywhere.
One test in plant_hydraulics_test.jl only checks the case of when rooting_depth is a float
because changing to a field requires extensive changes to the entire test.