Change plant hydraulics integration test to use fields
The test solves a system of functions which is defined in
initial_compute_expected_tendecy!. That function is now changed
to support fields, but each equation in the system of equations still
outputs a scalar. This is done by taking a mean across the resulting field.
This is done because NLsolve does not work of equations on fields.