Extend alpha and tau in radiative params to support fields
α_PAR_leaf, τ_PAR_leaf, α_NIR_leaf, and τ_NIR_leaf in TwoStreamParameters
support fields. α_PAR_leaf and α_NIR_leaf in BeerLambarParameters also
support fields. Both constructors changed to match the new structs.
The tests that used TwoStructParameters and BeerLambartParameters
now test functionality when the parameters are a float, and then as a field
of constants. This is a straight forwards change for all tests except
test_two_stream.jl, where a workaround is used to  avoid broadcasting
isapprox over a field of floats. This results in an error because
it tries to convert bools to floats.