Change plant_hydraulics_test to follow previous behavior
Previously the Plant hydraulics model integration tests
were changed to work with rooting_depth as a field. In the test,
the solution of a system of equations is found using NLsolve.
The previous change redefined each equation in the system of equations
to create a field of values the equation evaluates to, and then takes
the mean of the field so the equation still results in a scalar.

To better follow the functionality of the previous test, this commit
reverts the system of equations to only work on a scalar rooting_depth.
The solution to the system of equations is returned as a vector of floats.
It is not possible to create a ClimaCore Field of vectors. Instead, the
solution is stored as a vector of fields.

To implement this, a function that returns a function representing
a system of equations is created. This results in a system of equations
for each cell. At first, each system of equations was solved once for each
variable in the vector of solutions, resulting in many repeated equations solves.
To prevent this a dictionary is used. Each variable in the solution must be
returned independently in order to create a vector of fields.