Extend G_functions to support fields
The two G_function structs, CLMG and Constant G function both
hold a parameters. This parameter can now be a float or a
field of floats. To do this, a variable G was added to the Radiation
model's auxilliary vars. G is now calculated only once during
update_aux!. Previously, G was calculated once in extinction_coeff
and once  in plant_absorbed_pfd, which are both called during
update_aux.

G was added to the auxilliary vars because it is not possible
to broadcast on a struct containing a field. This means G needs
to be calculated before extinction_coeff and plant_absorbed_pfd
are called, and then passed in as an argument.

The function compute_G