Make changes non-breaking (again)
Added support for root_distribution. The PlantHydraulicsParameters
struct has root_distribtion again. Both root_distribution and
root_depth can be nothing. The constructor uses depwarn when
root_distribution is passed in. If neither root_depth or root_distribution
are passed in to the constructor, an error is thrown.

There were two uses of root_distribution. At these spots, a new
function is defined, which takes in the rooting_depth and a z as an argument.
If that rooting_depth argument is nothing, then it calls root_distribution
form the params. Otherwise it calls from src.

Removes breaking change from NEWS.md