Calibrate global model
This PR goal is to build a global calibration pipeline (minimal working example - MWE) to calibrate a land model globally.
In this commit, we calibrate a bucket model (the framework for full land model would be the same).
Two different approaches are added to this commit: the same calibration is done with just EnsembleKalmanProcesses.jl - EKP, or with ClimaCalibrate.jl (ClimaCalibrate.jl uses EKP, but has different backend, optimised for different HPCs).
In our MWE, we use ERA5 latent heat - LH and sensible heat - SH flux as target observations for the calibration. We stack monthly observation at n (e.g., 50) locations on land.

[skip ci]