Revert plant_hydraulics_test to use scalar root_depth
When root_distribution was removed, it was replaced with
rooting_depth, which can be a field. The Plant hydraulics model integration tests
was changed to test rooting_depth as both a scalar and a field. In order to
make the broadcast work, a wrapper function was used around NLsolve. This works
fine with cpu, but errors on gpu because NLSolve does not return a bitstype.
This commit reverts the test implementation and only tests rooting_depth as
a field.