Change Plant hydraulics model integration tests to use fields
Plant hydraulics model integration tests now tests rooting_depth
as a field and as a float. The issue with this test and fields on the GPU
was the usage of NLSolve. This test first solves for the steady state of the hydraulics
system using NLSolve. Then, the solution is used to check that evaluating the
tendecy of the model also results in a steady state. Because the rooting_depth
field is constant, this commit uses NLsolve with a float of that constant.
Then, when the model's tendency is evaluated, the field is used.