Fix plant_hydraulics_test to work on GPU (#813)
* Revert plant_hydraulics_test to use scalar root_depth

When root_distribution was removed, it was replaced with
rooting_depth, which can be a field. The Plant hydraulics model integration tests
was changed to test rooting_depth as both a scalar and a field. In order to
make the broadcast work, a wrapper function was used around NLsolve. This works
fine with cpu, but errors on gpu because NLSolve does not return a bitstype.
This commit reverts the test implementation and only tests rooting_depth as
a field.

* Change Plant hydraulics model integration tests to use fields

Plant hydraulics model integration tests now tests rooting_depth
as a field and as a float. The issue with this test and fields on the GPU
was the usage of NLSolve. This test first solves for the steady state of the hydraulics
system using NLSolve. Then, the solution is used to check that evaluating the
tendecy of the model also results in a steady state. Because the rooting_depth
field is constant, this commit uses NLsolve with a float of that constant.
Then, when the model's tendency is evaluated, the field is used.