Extend two stream parameters to space varying (#793)
* Extend alpha and tau in radiative params to support fields

α_PAR_leaf, τ_PAR_leaf, α_NIR_leaf, and τ_NIR_leaf in TwoStreamParameters
support fields. α_PAR_leaf and α_NIR_leaf in BeerLambarParameters also
support fields. Both constructors changed to match the new structs.
The tests that used TwoStructParameters and BeerLambartParameters
now test functionality when the parameters are a float, and then as a field
of constants. This is a straight forwards change for all tests except
test_two_stream.jl, where a workaround is used to  avoid broadcasting
isapprox over a field of floats. This results in an error because
it tries to convert bools to floats.

* Fix broadcast error

* Change experiments to use spatially varying  radiation params

benchmarks/land.jl, longruns/land.jl, longruns/land_regional.jl changed
to use α_PAR_leaf, τ_PAR_leaf, α_NIR_leaf, and τ_NIR_leaf from clm_data

* Make alpha and tau params same type

Now enforce that alpha and tau radiation params
are either all fields or all floats

* Extend G_functions to support fields

The two G_function structs, CLMG and Constant G function both
hold a parameters. This parameter can now be a float or a
field of floats. To do this, a variable G was added to the Radiation
model's auxilliary vars. G is now calculated only once during
update_aux!. Previously, G was calculated once in extinction_coeff
and once  in plant_absorbed_pfd, which are both called during
update_aux.

G was added to the auxilliary vars because it is not possible
to broadcast on a struct containing a field. This means G needs
to be calculated before extinction_coeff and plant_absorbed_pfd
are called, and then passed in as an argument.

The function compute_G

* Add varying xl to long runs

* Remove type annotations as suggested

* Move K to radiation aux vars