Bump minimum version of packages for 1.11
Some packages are not compatible with Julia 1.10 and the downgrade test
fails. The easiest way out is bumping their compat. This is not the best
solution (e.g., ClimaLand is compatible with CUDA 5.2 on julia 1.10),
but the real solution would be to go through upstream packages and ask
them to amend the registry, which is a lot of work for little benefit.