Add support for checkpoints
This commit adds support for checkpoint and restarting a ClimaLand
simulation. The functionality is tested in a bucket experiment, where I
verify that saving a simulation to disk and restarting it leads to the
same state as running the simulation in one go.

In the process, I had to bump some packages (because I am using the
latest version of ClimaUtilities). I could not get ClimaLand to stay
compatible with 1.9, so I bumped the minimum version to 1.10.

We should try to keep 1.10 as minimum version because it is LTS.