Change soil albedo to depend of soil moisture
PAR and NIR albedo are moved from the EnergyHydrologyParameters
to the auxiliary vars via AtmosDrivenFluxBC. The parameters now
hold PAR/NIR wet and dry as a field of float. The old method
of creating the parameters with PAR/NIR albedo is marked as deprecated,
and the constructor just sets both the wet and dry values to the
passed albedo.