longruns global seasonal cycle
We currently plot global maps for initial and final time steps (months)
of the long runs. In this commit, we add a global seasonal plot.
This plot takes the monthly global average of each output variables
and plot them as a line, the simulation is one year long so there is
12 data points.