longruns global seasonal cycle
We currently plot global maps of the long runs.
In this commit, we add a global seasonal plot for individual and all
years.
This plot takes the monthly global average of each output variables
and plot them as a line. Individual year in thin grey, average of all
years in thick black.