Fix warnings in tests
Replace deprecated reference_date arg

The ClimaUtilities DataHandler, TimeVaryingInput, and EveryCalendarDtSchedule
function were called with reference_date as an argument, which is
deprecated. This commit replaces them with start_date, and replaces
other uses of reference_date as an arg to functions that pass the arg
to EveryCalendarDtSchedule.

Fix warning in test for unused type declaration

Update common_diagnostics example in docs

Annotate start date