Tr/fix test warnings (#925)
* Fix warnings in tests

Replace deprecated reference_date arg

The ClimaUtilities DataHandler, TimeVaryingInput, and EveryCalendarDtSchedule
function were called with reference_date as an argument, which is
deprecated. This commit replaces them with start_date, and replaces
other uses of reference_date as an arg to functions that pass the arg
to EveryCalendarDtSchedule.

Fix warning in test for unused type declaration

Update common_diagnostics example in docs

Annotate start date

* Bump ClimaUtilities compat

ClimaUtilities compat bumped from 0.1.16 to 0.1.17.
This allows call to detect_restart_file to allways use
the non-deprecated argument for style.