add wrapper struct for type inference
This is required as a workaround to the type inference failure detailed
in https://github.com/CliMA/ClimaCore.jl/issues/2065. The failure appears
when a function is broadcast over one field and two structs; the types
are too complex and inference fails. This workaround introduces a wrapper
struct to contain the two struct inputs, so that the function is broadcast
over one field and only one struct, and the types are easier to infer.