Reduce allocations
This commit reduces allocations by changing turbulent_fluxes and net_radiation
to be in place. In addition to this, the workaround introduced to enable
GPU-compatibility in energy_hydrology resulted in lots of additional
allocations, so I modified that too.

There are still allocations. One of the easy ones to tackle is in
surface_specificy_humidity for snow, which accounts for a large fraction of the
total allocations for the snowy_land benchmark