Add and rename clm data artifact
Adds a higher resolution clm data artifact, named
clm_data_0.125x0.125. The previous artifact is renamed
to clm_data_0.9x1.25.

The clm_data_folder_path function takes in the `lowres`
flag as a kwarg, and returns the correct folder based on
that.

This also adds the `lowres` kwarg to the
`clm_canopy_parameters` and `default_spatially_varying_soil_parameters`
functions. The default value of `lowres` is determined by the
`use_lowres_clm` function. If given a cubed sphere mesh, the function
compares the results of ClimaCore.Meshes.element_horizontal_length_scale
to the approximate horizontal scale of both clm data artifacts
in meters. If given a rectilinear mesh, it does the same but in
degrees.

adjust test

use node horizontal length scale