Dashboard of land calibration(s)
This commit adds a code that produces a web Dashboard
to explore the results of a land calibration.
It allows to interact with all data available in a eki.jld2
object.
For example, plot of g, y and gamma for any variable, iteration,
ensemble member, season. Along with their parameters, error, etc.