Change `ALL_DIAGNOSTICS` to contain tuple
When a simulation is re-run by re-including the experiment script,
many diagnostic overwrite warnings are logged

The many warnings occur when `default_diagnostics` is called,
which repeatedly calls `add_diagnostic_variable!`.

This commit adds `default_diagnostics` as a kwarg to
`add_diagnostic_variable!`, with a default value of false.

This adds the kwarg with a true value to
each call to `add_diagnostic_variable!` in `default_diagnostics`

This also changes `ALL_DIAGNOSTICS` to have keys that are tuples
that contain the diagnostic variable and a boolean.
The boolean indicates if the
diagnostic is a default diagnostic or not.

Finally, `add_diagnostic_variable!` is changed
to not issue an overwrite warning if overwriting
a default diagnostic with a default diagnostic.