Make npolynomial = 0 by default
We still had some experiments and tests with non-zero `npolynomial`.
Diagnostics with the land-sea mask are correct only when `npolynomial`
is zero, so this commit ensures that all our tests and experiments are
using `npolynomial` of 0. This is how we want to run `ClimaLand` in
general.

I left one case with `npolynomial = 3` to check that it is still
working.