Restore support for ClimaCore < 0.14.30
ClimaCore `0.14.30` introduced support for masks and ClimaLand now uses
it by default. This commit restores compatibility with previous versions
of ClimaCore restoring the behavior before we had a mask: if users run
with previous versions of ClimaCore, they will not get the mask.

This compatibility allows us to replicate previous results.