Grab an extra year of data for era5 and modis LAI
It occured to a user wanting to make a 2 year simulation
from December to December, that the last December did not have
data for LAI. This was due to the latest LAI data point being in
November, and not extrapolated because the next year of LAI data was not
loaded. To consistently fix this issue, we add a year of data to be read
to interpolate the last month with the first month of the next year.
We also did this for era5 data file, where this error could also occur.