Use ClimaLand.landsea_mask for calibration
Previously to this commit, we were using `ClimaAnalysis.apply_oceanmask`
to get locations (coordinates) on land to perform global calibration.
This was causing erroneous values at the coast, because it would
sometimes use ocean locations, as the ClimaAnalysis mask slightly differ
from the ClimaLand mask. Furthermore, we use a high threshold (0.99) for
the land mask to ensure only land pixels are used.