Compute zenith angles for CoupledRadiativeFluxes
`p.drivers.cos\thetas` is currently updated by the coupler
when using `CoupledRadiativeFluxes`.

This commit updates the value in `update_drivers!` instead.
To do this, the commit changes `CoupledRadiativeFluxes`
to contain a function that calculates zenith angle in radians
given time from sim start date, longitude, and latitude.

This adds an additional constructor for `CoupledRadiativeFluxes`,
which takes in a start date (and optionally insolation params), and
returns a struct with a zenith calculation function that uses
Insolation.jl to compute zenith angle.