Make is easier to run a global calibration
In this commit, a few lines at the top of calibrate_land.jl code
are added and can be modified by users to design and run a calibration
job.
Previously to this commit, those lines where buried deeper in the code.
This makes it very easy and flexible to setup a calibration,
and sets up the direction to add further options (like, for example,
duration of calibration, noise configuration...)