Add GPU support to NeuralDepthModel
This commit allows NeuralDepthModel to be used
with a simulation on the GPU.

- Changes snowmip_simulation.jl to move arrays to cpu memory before plotting
- Adds Adapt.jl as a dependency. This is used to convert all arrays
in the density model into static arrays. This could be done without Adapt.jl,
but it would require far more code, because adapt can use
Flux's internal adapt_structure defintions.
- adds buildkite gpu run of snowmip

Note that this does not add cuDNN to the buildkite project.
cuDNN is required for full Flux gpu support, but the current
implementation of the NeualDepthModel evaluates the model inside
a broadcasted statement, which prevents the usual gpu optimizations
from batch evaluation from occuring.