Add flat noise weighted by lat to calibration
This commit adds an option at the top level of calibration
script to set a flat noise for the error covariance matrix
(diagonal). This flat noise is weighted by latitudes, so that
poles have higher noise, to avoid overfitting them as they
have larger area with the default grid setup.

to do: training_location in calibrate_land.jl is currently set
to Antartica only. Remove this before merging this PR.

[skip ci]