Remove nonfunctional `update_photosynthesis!`
This version of the P-model cannot be natively integrated within the canopy model yet, because it does not have support for subdaily updates of GPP. Thus we remove `update_photosynthesis!` in `pmodel.jl`. See #1206 for that implementation.

In the future, `compute_full_pmodel_outputs` should be used as a ClimaLand implementation of `rpmodel` that produces the full outputs.